+++ controllers/management_cluster_network.go
@@ -12,6 +12,7 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/predicate"
 
 	"github.com/aws-resolver-rules-operator/pkg/conditions"
 	"github.com/aws-resolver-rules-operator/pkg/resolver"
@@ -44,6 +45,11 @@ func NewManagementClusterTransitGateway(
 func (r *ManagementClusterNetworkReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&capa.AWSCluster{}).
+		WithEventFilter(
+			predicate.Funcs{
+				UpdateFunc: predicateToFilterAWSClusterResourceVersionChanges,
+			},
+		).
 		Named("mc-transit-gateway").
 		Complete(r)
 }
@@ -119,7 +125,7 @@ func (r *ManagementClusterNetworkReconciler) reconcileNormal(ctx context.Context
 func (r *ManagementClusterNetworkReconciler) applyTransitGateway(ctx context.Context, cluster *capa.AWSCluster) error {
 	logger := log.FromContext(ctx)
 
-	id, err := r.transitGateways.Apply(ctx, cluster.Name)
+	id, err := r.transitGateways.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 	if err != nil {
 		logger.Error(err, "Failed to create transit gateway")
 		return errors.WithStack(err)
@@ -139,7 +145,7 @@ func (r *ManagementClusterNetworkReconciler) applyTransitGateway(ctx context.Con
 func (r *ManagementClusterNetworkReconciler) applyPrefixList(ctx context.Context, cluster *capa.AWSCluster) error {
 	logger := log.FromContext(ctx)
 
-	id, err := r.prefixLists.Apply(ctx, cluster.Name)
+	id, err := r.prefixLists.Apply(ctx, cluster.Name, cluster.Spec.AdditionalTags)
 	if err != nil {
 		logger.Error(err, "Failed to create prefix list")
 		return errors.WithStack(err)