+++ CHANGELOG.md
@@ -11,12 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `RouteReconciler".
 
+## [0.12.0] - 2023-11-02
+
 ### Changed
 
 - Set user-agent to `giantswarm-capa-operator` when making requests to the k8s API.
+- Most controllers will skip reconciliation events that only changed the `resourceVersion` field of the CR.
 
 ### Added
 
+- Add tags from `AWSCluster.Spec.AdditionalTags` and `AWSManagedControlPlane.Spec.AdditionalTags` to  all created resources.
 - Add `global.podSecurityStandards.enforced` value for PSS migration.
 - Add managing prefix lists for management clusters and rename `ManagementClusterTransitGatewayReconciler` to `ManagementClusterNetworkReconciler`.
 
@@ -152,7 +156,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - changed: `app.giantswarm.io` label group was changed to `application.giantswarm.io`
 
-[Unreleased]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.9.0...v0.10.0